+++ NEWS.md
@@ -1,5 +1,7 @@
 # rticles 0.21
 
+- New `jedm_article()` for the Journal of Educational Data Mining template (thanks, @jooyoungseo, #251).
+
 - New `glossa_article()` for articles of [Glossa: a journal of general linguistics](https://www.glossa-journal.org/) (thanks, @stefanocoretta, #361).
 
 - All templates have now the `$highlighting-macros$` variables required for Pandoc highlighting (#435).